pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "latsym"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
]


[tool.setuptools]
package-dir = { latsym = "src" }
packages = ["latsym", "latsym.groups", "latsym.metrics", "latsym.tasks"]
